requirements.txt
aiohttp==3.11.13
asyncpg==0.30.0
pypika-tortoise==0.5.0
python-dotenv==1.0.1
tortoise-orm==0.24.2
uvloop==0.21.0; sys_platform != 'win32'
